refactor(cli): keep interactive deps out of the Node API bundle
Importing @react-grab/cli/api dragged in the interactive promptSkillInstall
(and its prompts/ora deps) because it lived in install-skill.ts, which the
API re-exports for installSkill/removeSkill.

- Move promptSkillInstall into utils/prompt-skill-install.ts; install-skill.ts
  now only carries the lean installSkill/removeSkill + light deps
- Point the init/add commands at the new module
- Align installReactGrab's pending-change predicate with init's invariant
  (originalContent + newContent + !noChanges) to prevent drift
- Fix the README low-level example to guard writes on !noChanges and reuse
  getPackagesToInstall()
